+++ notolog/ui/search_form.py
@@ -18,7 +18,7 @@ For detailed instructions and project information, please see the repository's R
 """
 
 from PySide6.QtCore import Qt, QSize, Signal
-from PySide6.QtWidgets import QWidget, QLabel, QLineEdit, QPushButton, QCheckBox, QHBoxLayout, QSizePolicy
+from PySide6.QtWidgets import QWidget, QLabel, QLineEdit, QPushButton, QHBoxLayout, QSizePolicy
 from PySide6.QtGui import QColor
 
 from . import Settings
@@ -26,7 +26,7 @@ from . import AppConfig
 from . import Lexemes
 from . import ThemeHelper
 
-from typing import TYPE_CHECKING, Union  # noqa: F401
+from typing import TYPE_CHECKING, Any, List, Dict, Union  # noqa: F401
 
 import logging
 
@@ -44,9 +44,12 @@ class SearchForm(QWidget):
     # Signals to mimic QLineEdit's textChanged and returnPressed events.
     textChanged = Signal(str)
     returnPressed = Signal()
+    searchButtonClear = Signal()
+    searchButtonNext = Signal()
+    searchButtonPrev = Signal()
     caseSensitive = Signal(Qt.CheckState)
 
-    def __init__(self, parent, search_buttons: list):
+    def __init__(self, parent):
         super().__init__()
 
         self.parent = parent  # type: Union[QWidget, ToolBar]
@@ -60,8 +63,6 @@ class SearchForm(QWidget):
         self.logging = AppConfig().get_logging()
         self.debug = AppConfig().get_debug()
 
-        self.search_buttons = search_buttons
-
         self.settings = Settings()
 
         self.theme_helper = ThemeHelper()
@@ -70,35 +71,60 @@ class SearchForm(QWidget):
 
         # Placeholder attributes for component widgets, initialized in init_ui.
         self._search_input = None  # type: Union[QLineEdit, None]
-        self._search_input_label = None  # type: Union[QLabel, None]
         self._search_pos_label = None  # type: Union[QLabel, None]
         self._search_count_label = None  # type: Union[QLabel, None]
 
+        self.case_sensitive_state = False
         # Search navigation button variables will be set via mapping
-        self.search_case_sensitive = None  # type: Union[QCheckBox, None]
 
         self.init_ui()
+
+    def get_toolbar_search_buttons(self) -> List[Dict[str, Any]]:
+        """
+        Retrieves the configuration map for search-related toolbar buttons.
+
+        Note: The 'var_name' parameter will be initialized in the final object that utilizes this map.
+
+        Returns:
+            List[Dict[str, Any]]: A list of dictionaries detailing the configuration of each toolbar search button.
+        """
+
+        return [
+            {'type': 'action', 'name': 'search_clear', 'system_icon': 'window-close', 'theme_icon': 'x-circle-fill.svg',
+             'action': self.searchButtonClear.emit, 'enabled': False, 'default': False,
+             'tooltip': self.lexemes.get('search_buttons_label_clear', scope='toolbar'),
+             'accessible_name': self.lexemes.get('search_buttons_accessible_name_clear', scope='toolbar'),
+             'var_name': 'btn_search_clear', 'color': self.theme_helper.get_color('toolbar_search_button_clear')},
+            {'type': 'action', 'name': 'search_prev', 'system_icon': 'go-up', 'theme_icon': 'caret-up-fill.svg',
+             'action': self.searchButtonPrev.emit, 'enabled': False, 'default': False,
+             'tooltip': self.lexemes.get('search_buttons_label_prev', scope='toolbar'),
+             'accessible_name': self.lexemes.get('search_buttons_accessible_name_prev', scope='toolbar'),
+             'var_name': 'btn_search_prev', 'color': self.theme_helper.get_color('toolbar_search_button_prev')},
+            {'type': 'action', 'name': 'search_next', 'system_icon': 'go-down', 'theme_icon': 'caret-down-fill.svg',
+             'action': self.searchButtonNext.emit, 'enabled': False, 'default': True,
+             'tooltip': self.lexemes.get('search_buttons_label_next', scope='toolbar'),
+             'accessible_name': self.lexemes.get('search_buttons_accessible_name_next', scope='toolbar'),
+             'var_name': 'btn_search_next', 'color': self.theme_helper.get_color('toolbar_search_button_next')},
+            {'type': 'action', 'name': 'search_case_sensitive', 'system_icon': '', 'theme_icon': None,
+             'var_name': 'btn_search_case_sensitive', 'enabled': True, 'default': False, 'color': None,
+             'tooltip': self.lexemes.get('search_case_sensitive_tooltip', scope='toolbar'),
+             'accessible_name': self.lexemes.get('search_case_sensitive_accessible_name', scope='toolbar'),
+             'text': 'aA',  # the 'aA' notation is commonly used universally to depict case sensitivity
+             'action': self.case_sensitive_toggle},
+        ]
 
     def init_ui(self):
         # Set up the main layout and internal components
         search_layout = QHBoxLayout(self)
-        search_layout.setContentsMargins(0, 0, 0, 0)
+        search_layout.setContentsMargins(0, 0, 5, 0)  # To allow spacing from the right
         search_layout.setSpacing(0)
-
-        self._search_input_label = QLabel(self)
-        # Apply font from the main window to the widget
-        self._search_input_label.sizeHint()
-        self._search_input_label.setText(self.lexemes.get('search_input_label'))
-        self._search_input_label.setObjectName('search_input_label')  # To differentiate it at styles file
-
-        search_layout.addWidget(self._search_input_label)
 
         # Searched text
         self._search_input = QLineEdit(self)
         self._search_input.setObjectName('search_input')
         self._search_input.sizeHint()
         self._search_input.setReadOnly(False)
-        self._search_input.setMaxLength(128)
+        self._search_input.setMaxLength(512)
         self._search_input.setPlaceholderText(self.lexemes.get('search_input_placeholder_text'))
         self._search_input.setAccessibleDescription(self.lexemes.get('search_input_accessible_description'))
 
@@ -128,25 +154,74 @@ class SearchForm(QWidget):
         self._search_pos_label.raise_()
         self._search_count_label.raise_()
 
-        for search_button in self.search_buttons:
+        for search_button in self.get_toolbar_search_buttons():
             if search_button['type'] == 'action':
                 # Create and append toolbar button
                 self.append_toolbar_button(button_conf=search_button, layout=search_layout)
 
-        search_case_sensitive_label = QLabel(self)
-        search_case_sensitive_label.setObjectName("search_case_sensitive_label")  # To differentiate it at styles file
-        search_case_sensitive_label.sizeHint()
-        search_case_sensitive_label.setText(self.lexemes.get('search_case_sensitive_label'))
-        search_layout.addWidget(search_case_sensitive_label)
+        # Connect to the event that clears the search form
+        self.searchButtonClear.connect(self.action_search_clear)
 
-        self.search_case_sensitive = QCheckBox(self)
-        self.search_case_sensitive.setObjectName("search_case_sensitive")  # To differentiate it at styles file
-        self.search_case_sensitive.sizeHint()
-        self.search_case_sensitive.setToolTip(self.lexemes.get('search_case_sensitive_tooltip'))
-        self.search_case_sensitive.setCheckState(Qt.CheckState.Unchecked)
-        self.search_case_sensitive.stateChanged.connect(self.emit_case_sensitive_changed)
+    def action_search_clear(self) -> None:
+        """
+        Clears the search field, resets case sensitivity, and updates the interface
+        to reflect the cleared state.
+        """
 
-        search_layout.addWidget(self.search_case_sensitive)
+        # Clears the search field by setting the current search text to an empty string.
+        self.set_text('')
+
+        # Reset case sensitivity.
+        # Unused: This line is commented out to allow independent handling of the button's toggle state.
+        # self.set_case_sensitive(False)
+
+        # Refresh the state of search buttons
+        self.action_search_buttons_refresh()
+
+        # Hide the occurrences counter
+        self.set_counter_text('')
+
+    def action_search_buttons_refresh(self):
+        """
+        Updates the enable/disable state of the search buttons based on the presence of search text and occurrences.
+        """
+
+        search_occurrences = self.counter_text()
+
+        # Retrieve the current search text.
+        text = self.text()
+        if hasattr(self, 'btn_search_clear'):
+            self.btn_search_clear.setEnabled(True if len(text) > 0 else False)
+
+        # Determine the state for navigation buttons based on occurrences.
+        btn_state = bool(search_occurrences)
+        if hasattr(self, 'btn_search_next'):
+            self.btn_search_next.setEnabled(btn_state)
+        if hasattr(self, 'btn_search_prev'):
+            self.btn_search_prev.setEnabled(btn_state)
+
+        # Highlight the case-sensitive button if that mode is active.
+        if hasattr(self, 'btn_search_case_sensitive'):
+            self.search_button_highlight(button=self.btn_search_case_sensitive, state=self.case_sensitive())
+
+    def case_sensitive_toggle(self):
+        """
+        This method switches the state of case sensitivity and notifies other components
+        of this change by emitting a caseSensitive signal with the current state.
+        """
+
+        # Toggle the case sensitivity state
+        self.case_sensitive_state ^= True
+
+        # Emit the caseSensitive signal with the current state to notify other components
+        self.caseSensitive.emit(self.case_sensitive_state)
+
+    def search_button_highlight(self, button: QPushButton, state: bool = False):
+        if state:
+            button.setStyleSheet("QPushButton { background: %s; }"
+                                 % self.theme_helper.get_color('toolbar_search_button_highlight', True))
+        else:
+            button.setStyleSheet("QPushButton { }")
 
     def append_toolbar_button(self, button_conf: dict, layout: QHBoxLayout) -> None:
         """
@@ -163,21 +238,37 @@ class SearchForm(QWidget):
         theme_icon = button_conf['theme_icon'] if 'theme_icon' in button_conf else None
         theme_icon_color = QColor(button_conf['color']) if 'color' in button_conf \
             else self.theme_helper.get_color('toolbar_icon_color_default')
-        # Theme icon with a fallback to a system one
-        icon = self.theme_helper.get_icon(theme_icon=theme_icon, system_icon=system_icon, color=theme_icon_color)
 
-        # Toolbar button with an icon
+        icon = None
+        if theme_icon:
+            # Get the theme icon, using a system default as a fallback.
+            icon = self.theme_helper.get_icon(theme_icon=theme_icon, system_icon=system_icon, color=theme_icon_color)
+
+        # Initialize a toolbar button with an icon.
         icon_button = QPushButton(self)
+        if 'name' in button_conf:
+            icon_button.setObjectName(button_conf['name'])  # Set object name from configuration.
         icon_button.setFont(self.font())
-        # Set the size (height) similar to search input field height maintaining the ratio
-        size_hint = self._search_input.sizeHint()  # As a real height hint check the line edit height
-        icon_button.setFixedSize(QSize(size_hint.height(), size_hint.height()))
-        icon_height = int(size_hint.height() * 0.7)
-        icon_button.setIconSize(QSize(icon_height, icon_height))
-        icon_button.setIcon(icon)
+
+        # Match button size to the search input field's height, maintaining aspect ratio.
+        size_hint = self._search_input.sizeHint()
+        icon_button.setMinimumSize(QSize(size_hint.height(), size_hint.height()))
+
+        # Configure button to display either an icon or text based on configuration.
+        if icon:
+            # Calculate and set icon size based on button height.
+            icon_width = icon_height = int(size_hint.height() * 0.7)
+            icon_button.setIconSize(QSize(icon_width, icon_height))
+            icon_button.setIcon(icon)  # Set the icon.
+        elif 'text' in button_conf and button_conf['text']:
+            # Adjust and set the button's font size.
+            icon_button_font = icon_button.font()
+            icon_button_font.setPointSizeF(icon_button_font.pointSize() * 0.8)
+            icon_button.setFont(icon_button_font)
+            icon_button.setText(button_conf['text'])  # Set button text.
 
         action = button_conf['action'] if 'action' in button_conf else None  # Action when clicked
-        if action is not None:
+        if action is not None and callable(action):
             icon_button.clicked.connect(action)
         if 'accessible_name' in button_conf:
             icon_button.setAccessibleName(button_conf['accessible_name'])
@@ -201,12 +292,6 @@ class SearchForm(QWidget):
         Emit the custom textChanged signal with the current text as the argument to notify of text changes.
         """
         self.textChanged.emit(text)
-
-    def emit_case_sensitive_changed(self, state):
-        """
-        Emit the custom textChanged signal with the current text as the argument to notify of text changes.
-        """
-        self.caseSensitive.emit(state)
 
     def text(self):
         """
@@ -236,7 +321,8 @@ class SearchForm(QWidget):
         Returns:
             bool: True if the case-sensitive search is enabled, False otherwise.
         """
-        return self.search_case_sensitive.isChecked()
+
+        return self.case_sensitive_state
 
     def set_case_sensitive(self, checked: bool = False):
         """
@@ -248,10 +334,8 @@ class SearchForm(QWidget):
         Returns:
             None: This method does not return a value.
         """
-        if checked:
-            self.search_case_sensitive.setCheckState(Qt.CheckState.Checked)
-        else:
-            self.search_case_sensitive.setCheckState(Qt.CheckState.Unchecked)
+
+        self.case_sensitive_state = True if checked else False
 
     def set_focus(self):
         self._search_input.setFocus()
@@ -282,6 +366,8 @@ class SearchForm(QWidget):
         self._search_count_label.setText(count)
         self.adjust_occurrence_label_size(label=self._search_count_label)
         self.set_position_text('')
+        # Refresh the state of search buttons based on current search conditions.
+        self.action_search_buttons_refresh()
 
     def position_text(self):
         """
@@ -332,8 +418,7 @@ class SearchForm(QWidget):
         label.resize(space, self._search_input.height())
 
         # Adjust the label's size and position dynamically when the widget is resized
-        label.move(self._search_input_label.width() + self._search_input.width() - label.width() - shift,
-                   self._search_input.y())
+        label.move(self._search_input.width() - label.width() - shift, self._search_input.y())
         self._search_pos_label.resize(label.width(), self._search_input.height())
 
     def resizeEvent(self, event):